Switched to Pyrex 0.9.2.1.
Pyrex is now tracked in http://codespeak.net/svn/vendor/Pyrex/.
It is copied from there and modified as needed for PyPy.
The original files are:

A    Pyrex
       http://codespeak.net/svn/vendor/Pyrex/Pyrex-0.9.2.1/Pyrex/

A    pypy/tool/pyrexc
       http://codespeak.net/svn/vendor/Pyrex/Pyrex-0.9.2.1/bin/pyrexc

The external links to Pyrex and Plex have been removed.
Note that Plex is a subdirectory of Pyrex in this new version.

I guess you will have to delete your Plex directory manually; removing the
svn:externals isn't enough for svn up to delete it for you.