Finished a rewrite started some time ago:  the dot graphs are now rendered entierely by a Python script from a description produced by 'dot' in its so-called 'plain' format, which includes text, coordinates, and bezier control points for the edges.
This allows much more flexibility.  Try graphviewer.py.  Variable names are
highlighted; the mouse position can be matched to the text more precisely than
was possible with the previous hack; the rendering's scale can be changed
easily (try dragging the right mouse button).

The goal is to display very large graphs with a reasonable amount of RAM
(before, large graphs were large bitmap images).  Also, the zoom feature will
be convenient for that.