All tests pass!
Argh, method support was non-obvious.  There is an "XXX messy", too.  But the
idea is clean: a CMethod representation represents a callable together with a
PyObject*.  Calling it is equivalent to calling the underlying function with
an extra first argument of type R_INSTANCE.  We rely on the optimization of
LLOp.optimize() to avoid copying around the 'self' argument all the time.

This is equivalent to Psyco's virtualized objects; in Psyco a virtual method
object is one that is not actually built by the generated machine code, but
the compiler (Psyco or typer.py) remembers that there is a method object that
virtually exists, with a 'self' field still stored in some other register,
that doesn't have to be really copied physically.  (A machine register for
Psyco is the equivalent of a LLVar in typer.py.)