A long function with no associated test.  Ugh.
This is an attempt to solve the problems related to generating C code for
'list += otherlist' expressions.  This involves generating a loop with
potentially conversion operations inside.  Where should the loop come from?
It is tricky to generate its C code manually (e.g. writing a for(...) directly
in the C source from genc_op.py) because the copy-and-convert may fail, and
explicit control flow in the C code written by genc_op.py doesn't mix well
with typer.py's efforts to Py_DECREF things correctly after errors.  Indeed,
typer.py assumes that the control flow is exactly as specified by the control
flow graph (duh).

So this tries to attack the problem earlier: transform.py will detect the +=
operation and replace it with new control flow graph blocks that do, in a
loop, the getting-items-from-list2-and-appending-them-to-list1 dance.

This looks good, actually: for ease of code generation as well as for
optimizations, transforming the control flow graph in various ways will help.
Maybe we should have utility functions to make this easier...  Currently it's
pretty verbose to manipulate control flow graphs.

There is no test for transform.py.  But it's not obvious how to test it...

All that remain to make the tests pass again should be to support in genc
these new introduced SpaceOperations (growlist, fastappend).