Lots of small changes to lots of files: trying to get Python 2.2 to work again.  Some tests still fail.  Not sure if it's worth the effort...
Also contains a new checkgraph() that checks various invariants of flow
graphs, e.g. that variables are used consistently and never in more than one
block.  Graphs now have a method show() to view them using the pygame code,
which is handy to figure out what's really wrong with a graph when an
assertion in checkgraph() fails.