Extended the Arguments class to include a 'w_stararg' attribute, which can hold a non-unpacked *argument.
The main change for the flow graphs is the operation 'call_args' replacing the
'call' operation.  ('simple_call' is still there for simple cases.)  See the
description of 'call_args' at the beginning of translator.transform.

The goal was to make *arg unpacking lazy; the 'call_args' operation can now
specify explicit keywords (not built as a dict) as well as this unpacked *arg.
In other words, all argument information is stored into the Arguments class
and packed unmodified in the 'call_args' operation, without having to expand
the *arg.

This starts to look a bit nicer, particularly for translator.annrpython, where
a pagefull of code has been replaced by just a call to Arguments.parse().
It's nice to be able to use the same parsing logic in the annotation phase,
though it's still not 100% clean: the Arguments class depends on the presence
of a space, but we don't have any during annotation.  There is a pseudo-space
RPythonCallsSpace in annotation.unaryop, but it's a hack.  The Arguments class
should be rewritten to be more independent on any space.