updating and improving some RPython chapters, they often refered to something as being in the future than in the present.  I also introduced some links to source code and starkiller. feel free to review and further enhance.
I note that the "integer" chapter is both
somewhat outdated with respect to the new
ovfcheck() and is sometimes written in the 'I'
form.

Would be great if someone can state
some authoritate things about the current
integer situation and refactor the 'I' into
a 'we' and follow the style of our other
documentation some more.