A preliminary checkin of the python translation of the _codecsmodule (and friends).
The module _codecs exposes the same methods as _codecs.pyd in CPython. As in CPython it is a wrapper for the actual codecs which I have put in unicodecodec.py ( for now). Not all codecs are implemented yet, and not all test passes.

I had to change the test2/test_codeccallbacks.py in order to test the codecs without involving unicode strings.

There some serious problems to be solved :

1) how to make a global registry from appspace. I have used module variables (in _codecs.py) for now. Ther may be better ways.

2) UnicodeError,UnicodeEncodeError,UnicodeDecodeError exceptions ar faked ? doesnt work as in Cpython anyway.

3) In translating the design from _codecsmodule.c I have discoered a sideeffect. The Encodings package injects the functions exposed in _codecs.py into a class. This converts the function into an unbound method. In a comment in the Encodings package it is mentioned that this doesnt happen for c functions.

This is work in progress