issue53 testing
removed the need to manually add SVN/file references
as long as:

- you specify them relative to 'dist' or to 'pypy'

- linknames contain a slash

To avoid having to generate thousands of links without
them being used (docutils parsing is slow) the logic
for generating references is now reversed: it starts
off from the .txt files and generates all neccessary
links.  This means that if you add file references that
previously didn't exist (py.test fails then), you have to run

    python documentation/tool/makeref.py

Not sure yet, if this run should be integrated into py.test
because it really modifies files (_ref.txt) and that is uncommon
to do for testing.