issues16 in-progress
 changes make sense.

changed the wording again a bit about old-style classes, custom metaclass seemed to imply even more magic
that there is. Old style classes have of course their metaclass and we simply implemented it as user-code
(that it is geninterped is an impl detail).

Olny question: why call the last section "Disclaimer"???