issue65 in-progress
reviewed the list of references and found a lot of them
not really fitting anymore or not very helpful.  I thus
reduced the list which is basically now a better start of a
'related projects' sections.  We should probably invest some
more effort and make a real 'papers and talk' section
on the index page, referencing our so-far presentations
(including the OSCON one).

Before removing references i would have liked to talk
this over with someone but no one was available and
release time is coming close.