issue16 resolved
removing the somewhat strange reference to the
"codespeak.net umbrella" although the weather
is not nice right now and an umbrella wouldn't hurt
and i also wouldn't mind some free advertisement.
But it's a bit ouf ot context here (or am i missing
something?). Closing this issue.  If you want the
umbrella, reinsert it.