long objects are complete and pass all tests.
Karatsuba isn't done, yet.
Also we are currently stuck with 15 bits, because
long division needs a signed type that can hold two digits *and* a sign.

Temporarily, I had an extra class to emulate this, and all algorithms
are tested to work at full bit size. I removed the class because
I was not pleased with the overhead.