Experimental hook to have more fine control over which blocks get specialized in which order.
The order.py module is an example using it, it uses a list of modules names (or <module-name>:<func-name>) to sort the blocks

use RTYPERORDER=order and edit locally 'module-list.example' as 'module-list' to try it out.

with this approach I discovered e.g. that the ~8500 blocks from (synthetic) functions without module (module=None) all specialize fine.