Hey, marshal is here!
I added an interp-level marshal module. It is also enabled
to do the basic marshalling and loading of .pyc files.

This implementation is fairly fast. There is some performance
to be gathered in the presence of files. Buffering has not been
taken, yet.
Hint: PyPy marshal does not use file descriptors, but file-like
objects. For optimization, look for a seek method and redirect
to the string marshaller. Make sure to seek back to the true end
of the marshal. This is cheap, but make PyPy marshal much more
versatile than CPython's.

This implementation tries to be as compatible to builtin marshal as possible.
At the time being, there is one problem: Recursion depth is not big
enough to emulate CPython's nesting depth of 5000. I did a couple of
aproaches to do it the stackless way. The snippets can be found in
stackless_snippets, but nothing was pleasant enough, yet to make me
incorporate it into the so far quite nice module. After all, every
simpler effort brings me back to the way Stackless 3.1 implements things.
I will finally implement an extension this way.

Although I was not pleased with my temporary results, I have to say
that RPython's object model is much nicer than that of the C language!
The single inheritance is very, very powerful. I will use it to implement
custom frame-like structures. I also agree with Armin's comment that
such transformations should not be done by hand, but that we should
find algorithms which generate this automatically. Besides that, I think
it makes sense to explore these things, and marshal is a perfect example
which is not too complicated, but reveals a number of non-trivial cases.

Nevertheless, for complete CPython compatibility, it is necessary
to limit object nesting to 5000, or CPython will crash. I do intend
to back-translate the stackless version to CPython, in order to
remove this arbitrary limitation. I'm also expecting new random
arguments whioch will prevend this from getting into the core,
as usual :-)