follow up to r11953 to deal with multimethods attaching methods while annotating.
r11953 assumption was that this would result in a matching with no uplookup match,
but this is not the case if subclasses share a same impl function because this result
in matching receiving a unified PBC with the impl function attached to a superclass
classdef resulting in a possibly successfull uplookup match.

Double-check in that case that uplookup classdef was considered as possible source
for the impl function, otherwise look for possible missing definitions for the attribute.

Now annotation terminates with query.sanity_check_for_methods not displaying any problem.