first stab at integrating garbage collection in the llinterpreter. This is probably still quite flaky. At the moment only collectors that don't need read/write barriers are supported and collectors that need only the roots.
I cheat in some places: the offsets in objects to pointers to other
objects are found by using the searching through the lltype. this can
be computed in advance, but isn't at the moment.

This all is not very well tested yet :-(. I suspect that there are
some bugs. I need to check in to change computers.