(mwh, arigo, pedronis around)
Did stuff.  Reorganized a bit the classes: LLAbstractFrame is replaced by a
BlockBuilder, responsible for building only one block; the process of
completing a graph is now driven by the GraphState directly.  We also replaced
most usages of 'isinstance(_, LLXxxValue)' with method calls, in preparation
for the next test -- a new LLXxxValue class for virtual structures.