(arre, pedronis)
implemented proper new block creation logic in enter_block. merge logic is still not there,
just calling enter_block for now.

Refactored rgenop interface to take wrapped constant lltypes, for convenience writing
enter_block. Notice that constTYPE can be called insided rtyped functions so this is not
too bad.

Now the timeshifted functions return a jitstate with attached a curvale redbox (either
constant or with a variable in it)

everything is very much in flux.