(cfbolz, pedronis):
experimental change to the logic object space:

use == and cmp and is to unify logical variables. This is really only to get a
feeling for the issues but it's actually very cool, because you can unify lists,
tuples, dictionaries, and even class instances, if they have an __eq__ defined
that uses equality of the __dict__ of the instances.