Added a failing test.
Basically, if we modify the Record after its hash has already been
computed, we might obtain two objects that compare equal but hash
differently.

Disabling the caching in LowLevelType.__hash__ fix the test, but I
think it's not what we want.

Adding the following __hash__ to ootype.Record also fix the test, but
I'm not sure if it's correct, especially with recursive structures:

  def __hash__(self):
    return hash(self._fields)