(pedronis, arigo)
An extension to extregistry: you can pass a policy object to the various
lookup functions, and register entries with conditions based on the
policy.  See test.  Purpose: special-casing rgenop when compiling to C
on a specific processor; but this could also be used e.g. to select
between lltypesystem and ootypesystem versions of built-in functions.