(arre, pedronis)
start of having the hint-annotator exception transform the copied graphs it annotates.

with this in place we need to be careful about what pointer comparisons mean for virtual or
red virtual containers. Not all tests work with exceptiontransform on, to have them pass
we need to avoid pointer equality comparisions to force things to eagerly: next thing to do.

(For completely blue container this is controlled at annotation time, see code
added to annotate ptr_eq/ptr_ne for constant/virtual container pairs).