(pedronis, arigo)
Pass an extra test about catching exceptions in JITted code.

More and more based on the exceptiontransform, this required some
clean-ups and a strange flag to the latter to produce graphs that give
more hints to the JIT.

Added a sanity-check to llinterp: when returning from
exceptiontransformed graphs with an exception set, the return value
should really be the special error value.