(pedronis, arre, arigo)
Merge the jit-promotion branch:

   svn merge http://codespeak.net/svn/pypy/dist@32682
              http://codespeak.net/svn/pypy/branch/jit-promotion

It adds "promotion": with  x = hint(x, promote=True)  we can force a red
variable to become green.  This is implemented by pausing the
compilation and resuming it only when we get a run-time value arriving.
The "pause" contains run-time code that is a "flexible switch": a switch
where new cases can be added dynamically.  The default case is the one
that calls back the compiler.