profile based (call counts) inlining: --prof-based-inline=<args to the program for profiling> There's even a test.
A first stab. Based on forking...

I managed to get an -O3 executable for PyPy that's is slightly smaller but not slower. No such luck with -O2.
Still kind of good considering our failures so far to reduce the inlining amount without killing the performance.

Didn't spend time on tuning or playing with the parameters at all.