(pedronis, arigo)
Intermediate check-in #2.

More seriously, we rewrote the interpreter main loop in a style that
should be easier for the JIT to grasp, but also -- possibly -- easier
for human beings to grasp.  A bit.  Many levels of wrapping of stuff
into other stuff are avoided now, and the PyFrame class hierarchy was
completely removed.  There is a possibly nice way to add opcodes from
outside, with config options to enable them.  The drawback is that the
core main loop uses modern RPython magic.

Translation not tested.  It makes py.py slower by 30% but we hope that
pypy-c will be faster.