umph, reverting the ootype malloc removal refactoring 36037 (and consequently 36040), it needs more careful review.
It broke in a shallow way backendopt/test/test_all.py but once the shallow breakage is repaired we see that
a bunch of tests there are failing because malloc removal is not happening anymore it seems.

This needs to be fixed in such way that test_all and translator/test/test_simplify pass again and then it can be
checked in again.

test_all is quite an important test for *all* backendopts!