Simplification of llmemory, triggered by the fact that two addresses would compare as different if they were obtained in different ways, even if they pointed to the same final object.
The 'offset' field of fakeaddress is gone, and 'ob' is renamed to 'ptr'.
It's always just a pointer.  With the new stuff in lltype like
direct_fieldptr() we can take pointers to fields, or array items, so no
need for anything more.

Hack a bit for arenas and address-of-length-of-array.

A new test in rctypesobject that triggered this whole thing,
and implementation of recursive data structures.

llvm/database.py updated but not tested (will do so on another machine).

gctransform/test/test_framework took ages because of the llinterpretation
of a raw_memclear() of the root stack, with 160'000 items!

Trying to do something cleaner about raw_memcopy() in llmemory.py.