sanitize a bit the ExternalType mess:
unify SomeExternalObject and SomeExternalBuiltin, introduce SomeExternalInstance which has just different union behavior.

use getrepr and introduce rbltnregistry for rtyping logic for BasicExternal whose annotation now is a special case
of SomeExternalInstance

split ExternalType into a thin OOType and a desc that is cached on the bookkeeper. No more mutable ootype containing
annotations ...

the js test are passing, as well as the annotation and rtyping tests for BasicExternal for what is worth.