add the possibility to export classes.
A class is exported is its __init__ is marked with @export. Individual
methods can also be @exported, useful for methods that can't be
reached by one of the entrypoints.

To force the annotator to analyze a given method, we generates a dummy
function which call that method. Better approached are welcome.