this is a replay of r62493 by afa
-----------
in oosupport/constant.py, the constants cache store objects of different
types in a dictionary: all keys should be comparable.

I was unlucky enough to have two objects with the same hash.
but this can be explained if one was a _view of the other.
Change a bit the __hash__ function to reduce collisions.
-----------