Merge the resoperation-refactoring branch.
The idea is to reduce the memory needed by ResOperations, by putting certain
fields (e.g. a variable-sized list of arguments, or descr, or fail_args) only
on the operations that actually need them.

The benchmarks don't show any speedup or slowdown, but it's still worth
merging because it saves a bit of memory: list of operations are thrown away
immediately right now, but we might want to keep them in the future.