(antocuni, arigo)
Merge branch/jit-free.

The major point of the branch is to reduce memory usage of long-running
pypy's with the JIT (such as "pypy translate.py").  Done by playing with
weakrefs and having a __del__ that kills the ResumeGuardDescrs of old
loops.  This is controlled by "--jit loop_longevity=N".  (Actually
freeing the assembler code would be the next step, but is less important,
because the ResumeDescrs take more memory.)

In addition, this branch contains a number of semi-related things:

    * Generators: set their 'frame' to None when they finish.  This
      helps a lot the GC.  (The issue is that generators need an
      interp-level __del__; this let the GC free the frame earlier.)

    * rgc.dump_rpy_heap() now works with almost no need for
      additional memory.

    * pypy.rlib.debug.debug_print(r_longlong(...)).

    * add debug_{start,stop,print} to the __pypy__ module.  This lets
      you write to the PYPYLOG file from app-level.

    * use this new interface in translator/driver, to record the begin
      and end of tasks (via pypy.tool.debug_print so that it does not
      crash on CPython).

    * improve a bit reftracker.py.

The net result is that you can now translate pypy with a pypy-jit on
32-bits, using roughly 2.5 GB of RAM.  It is twice as big (for now) as
CPython, but twice as fast :-)