Merge branch/rlist-jit.
Improve in jtransform.py the handling of constructs like
'frame.locals[n]'.  This is now done with the help of the
annotator: from the hint _immutable_fields_=['locals[*]']
found on the class, the annotator flags the list returned
by 'frame.locals' as being an immutable list; then the
rtyper translates 'frame.locals[n]' with a call to
ll_getitem_foldable instead of ll_getitem.

The goal is to make an extra test in test_list.py pass:
handling IndexErrors out of RPython-level lists.  This is
done essentially by letting the JIT see the code
"if n>=len: raise IndexError" in rlist.py.

Other changes (related):

* fixes left and right because _immutable_fields_=['lst[*]']
  now implies that we only read from the list using a known-
  nonnegative index (checked in jtransform.py)

* cpyext refactoring about tuples