(dmalcolm, lac): attempt to fix links to source code within sphinx-generated HTML
Sphinx generates the html in _build/html, which means that all the links relative to pypy/doc break within the html files.

The real fix appears to be to write a plugin to sphinx to handle this more gracefully, perhaps based on http://sphinx.pocoo.org/ext/viewcode.html

For now, this changeset merely tries to consolidate all links to source code files to be of the form ../../../../pypy

This was achieved using these sed commands:

  sed -i -e"s|: \.\./objspace/|: ../../../../pypy/objspace/|" *.rst
  sed -i -e"s|: \.\./\.\./pypy/|: ../../../../pypy/|" *.rst
  sed -i -e"s|\.\./\.\./pypy/|../../../../pypy/|" getting-started-dev.rst

to avoid accidentally modifying paths in other places (e.g. usage examples)