(prompted by amaury)
In CPython some codecs raise IndexError but others raise OverflowError
in the error handler if the integer position is out of bounds.
Decided that just sticking with OverflowError to make the code
simpler is a good enough solution.  Fixed a few tests in lib-python
to accept either IndexError or OverflowError.