Rewrite the logic to handle threads in the shadowstack root finder. Based now on custom_trace.  This should be better because it allows thread that block for a long time to not have their shadowstack scanned again and again at every minor collection.
It should also allow us to reduce virtual memory usage by putting
a limit on the number of full-blown shadowstacks around (useful
on 32 bits if you have many threads).

This last benefit is the key to this rewrite, which will make it
reasonable to use the same code for stacklets too (future work).