bah :-(
cpyext rely on some kind of global state which I could not sort out. As a
consequence, if you try to instantiate two different objspaces in the same
process, the second one explodes.

The only way to make it working is to make sure that *all* gettestobjspace
calls have they very same config: this way, there is a cache which reuses the
already-built objspace, so we don't build a second one and things work.