Fixed missing import of compute_unique_id.
Changeset e3249f3bf105 introduced several uses of compute_unique_id,
which was only imported inside dump_storage and not globally.

This lead to translation with -opt=jit on PyHaskell failed. This could
possibly affect other RPython VMs too.

The fix is simply to move the import of compute_unique_id from inside
dump_storage to the top of the file.