intermediate checkin: start removing resume_at_jump_descr
it was used in unroll.py as the target to jump to for newly created guards.
this was hugely complicated and I suspect actually wrong in some cases.
Instead, generate a dummy guard before a jump. This guard is removed by
optimizeopt, but its descr can be used by unroll when inventing new guards.

almost finished, but some tests show differences in the loop enter count, still
to be investigated.