Introduce promote_unicode().
This is a companion to promote_string() which promotes Unicode strings by
value. To use, simply ``from rpython.rlib.jit import promote_unicode`` and
then ``promote_unicode(any_unicode_string)``.

A few tests are included, and a couple bits of testing mocks were improved to
permit tests to not fail.