(wip) add handling for pinned objects referenced by old ones.
The added AddressStack keeps track of old objects which point to young
pinned objects. This way if a pinned object is only referenced from an
old object, it isn't collected as before.

This is a work in progress. For now test_object_pinning.py tests pass.
However, there is still a lot to be tested and double checked (see XXX).
In addition some refactoring & additional comments are in order after
we make sure it works as expected.

For now this changes stop pypy from being translated.