Stop tracing when inlining is detected.
Currently, if a function called part way through a trace turns out to be
recursively, it is endlessly inlined, often leading to a tracing abort. This is
tremendously inefficient, and many seemingly innocent recursive functions have
extremely bad performance characteristics.

This patch is the first step in trying to address this. Notice that this breaks
a few tests, in the sense that it changes what is traced. It should have no
visible effect on end behaviour.