hg backout e36875ce12ab
This was a bogus solution.  We can see in x86/runner.py that multiple
invalidate_loop() with the same looptoken never patches the same
location more than once.  We need to debug and understand the original
problem again.

Additionally, e36875ce12ab has a bad performance effect on some
benchmarks.  I think it is because it is possible to have one loop that
gets invalidated, but not the other loops that JMP or CALL_ASSEMBLER to
it.  That's how a guard_not_invalidated can fail a lot, and a reason why
attaching a bridge to it can be a very good idea.