Don't reraise CompilationError as ImportError in rpython.rlib.rzlib.
This was a bit annoying if rffi_platform.configure_external_library failed for
reasons other than missing zlib headers because it masked the actual error.  I
also changed all places which expected rzlib to raise ImportError.  If I missed
something or you disagree with this change, feel free to back it out.