Avoid import in app_main.py before stdio is initialized.
If the interpreter is run with the '-v' option specified, the import machinery
prints messages to sys.stderr.  This commit removes two imports before stdio
initialization, which resulted in an AttributeError because of this.

The imports were there to avoid issues during bootstrapping, which I did not
yet run into after removing the imports.