Remove this test because it doesn't make sense anymore.
In the default branch (implementing 2.7), the test checks that, even though
the script first writes to sys.stdout, stderr is output first.  This works
because both streams are buffered and, at interpreter shutdown, sys.stderr is
flushed first.

The order in which sys.stdout and sys.stderr are flushed isn't determistic if
running on Python 3.x.  Because of this, the test was changed in the py3k
branch (at the time trying to implement Python 3.2) to check that both streams
are always line buffered.  However this assumption is false.  Since Python 3.0
sys.stdout and sys.stderr are line buffered by default iff connected to a TTY
or set otherwise unbuffered.