consider freshly allocated memory in inevitable.py
As long as we are in the same transaction, freshly allocated memory (even raw)
will just get freed on abort. There is no way it could escape from the current
transaction before a successful commit, so we do not need to become inevitable
if we know that the memory was allocated in the same transaction.
inevitable.py was adapted to keep track of freshly allocated memory and to not
turn inevitable on writes to such locations.