Let the return type of direct_arrayitems() track whether it is a pointer into a GC object.
direct_arrayitems() returns an interior pointer into an object which might be a
GC object.  With STM, GC objects are in a special address space.  Interior
pointers into GC objects point into this address space, too.  This change makes
GenC emit the correct type for interior pointers into GC objects.

This is needed for clang, because it doesn't allow implicit pointer casts
across address spaces.  GCC doesn't need this because those implicit casts are
allowed and the pointer is casted to the correct type before accessing the
object.