slightly different approach for integrating mutable cells and immutability:
before, an integer field went from storing a W_IntObject immutably, then
switching to a IntMutableCell on the first mutation. This makes everything less
type stable, leading to more bridges and making this branch fundamentally
incompatible with heap profiling. Now the IntMutableCell is *always* stored,
but if it is ever mutated map.ever_mutated is set to True. That means that as
long as ever_mutated is False, it's safe to fold the read from the
IntMutableCell away as well, in an @elidable method.