Fix cpython test suite's SSL tests
They were failing since the expiration of certificate for svn.python.org
at Thu 24 Dec 2015. See details at cpython's Issue #25940. [1]

This patch is an import of cpython's patch fixing the issue, see [2].

The only modifications are in the paths that are different in pypy's
repository and the missing Misc/NEWS file for which the hunk has been
dropped.

[1] http://bugs.python.org/issue25940
[2] https://hg.python.org/cpython/rev/fb7131939508