Preimport module to reduce stack depth.
This test file passes if run in isolation.  However, if run in conjunction with
test_greenlet.py, it fails with a "maximum recursion depth exceeded"
RuntimeError because using greenlets untranslated on top of CPython results in
the recursion counter being off.  This turned out to be too hard too fix, so
instead I changed the subsequent test to use fewer stack frames.