Fix C API test.  See comment for details.
Thanks to Ronan and Armin for discussion and the suggestion to mask out bits.

It would also be possible to mask out bits in the Py_REFCNT macro.  There is a
discussion whether this should be done (also on default).  In this case this
changeset wouldn't be needed and should be backed out.