Compile with -Werror in cpyext tests.
Callers may build with -Werror and should not succeed with CPython but fail
with cpyext. This forces us to have the same API -- in particular, to use the
same pointer types and perform the same casts. See, for example, the change to
the API in ndarrayobject.py.


(This is an hg backout of changeset db306c552216 from when I accidentally
 pushed this to pypy without code review... hee hee. :C)