Fix translation somewhat. RPython can't handle constants that are overridden at class scope.
This still has the killer issue that mass_free_incremental is passed two
different callbacks, and so I get a type error when it tries to unify the two.

I think that this means that I can't use an ArenaCollection for storing the GC
headers.

TyperError: cannot find a unique name under which the methods can be found:
  [<MethodDesc '_free_if_unvisited' ...>,
   <MethodDesc '_IncrementalMiniMarkRemoteHeaderGC__free_flags_if_finalized' ...>]