fix the problem
when attaching a new bridge to a GuardCompatibleDescr and we cannot find out
whether the new bridge is checking conditions on the same variable as the
GuardCompatibleDescr, the guard needs to henceforth always jump to that new
trace if none of the other traces match. Otherwise we will keep compiling more
and more code.