protect setting the execution context threadlocal with inevitable
Seems to fix a crash where the threadlocal was corrupted. If not setting the TL
in an inevitable TX, the superclass' try_enter_thread would not re-set the EC
after an abort. Instead, it would detect that the TL already has a value and
probably reuse it in multiple threads.