Inline gotos to blocks with only one predecessor.
This is a small subset of the patch in issue #1220.

Example output:

Signed pypy_g_main(Signed l_x_0) {
        Signed l_v131;

        /* block0: (inlined) */
        OP_INT_ADD(l_x_0, 1L, l_v131);
        /* block1: (inlined) */
        RPY_DEBUG_RETURN();
        return l_v131;
}

Instead of something like:

Signed pypy_g_main(Signed l_x_0) {
        Signed l_v131;
        goto block0;

        block0:
        OP_INT_ADD(l_x_0, 1L, l_v131);
        goto block1;

        block1:
        RPY_DEBUG_RETURN();
        return l_v131;
}

The readability benefit is when you have if statements and such.

This does NOT handle a case like:

if (x) { <...>; goto block1; }
else { <...>; goto block1; }
block1:
...

That case is a bit more fragile and maybe not worth the cost of implementing.