Update int(x, base=10) behaviour for Python 3.4.
"Changed in version 3.4: If base is not an instance of int and the base object has a base.__index__ method, that method is called to obtain an integer for the base. Previous versions used base.__int__ instead of base.__index__." (from https://docs.python.org/3/library/functions.html#int)

Note that in CPython an error is returned if base.__index__ does not exist on the non-int object, even if base.__int__ exists. Test amended to confirm this error is raised.