Build main binaries on OpenBSD with wxallowed.
On RPython, JIT compilers that need to read and write to the same page need to
be marked as wxallowed (unless they've been buit to cope with this restriction).
Previously, this meant specifying LDFLAGS in the environment before building an
RPython VM which I always forgot to do. This commit automatically marks the
final binary as wxallowed without any such annoyances.