If the machine can be [3-6], allow the suffix to be [3-6].
It's hard to tell but the intent (https://www.python.org/dev/peps/pep-3149/) seems
to be to embed the machine identifier into the suffix.  Therefore, the old version of
the test was broken in asserting that i[4-6]86 machines all turn into i386 in the
suffix.  Change the test to accept the other machine identifiers as well.