Explicitly use Py_ssize_t as the Signed type in pypy c-api.
The signature for the c-api uses Py_ssize_t for the unsigned arguments.
`long` is equivalent to `Py_ssize_t` (in fact `Py_ssize_t` is a `typedef` of
`long`) in all cases in C, but when integrating with go, `C.long` and
`C.Py_ssize_t` are considered two disparate incompatible types.  By using
`Py_ssize_t` in the function signature, pypy + cpython compatible code can
be written.