(antocuni, arigo)
Hack around: add the @c_only decorator that makes it possible to write
a function in RPython that is callable (only) from C, with no magic.

Use it for free(): we want it to occur in RPython because of
lltype.malloc(), and the tests are not happy because ll2ctypes never
sees the free.