WIP: try a different approach for descr_call: instead of creating a different class/typedef for each calling convention, use a very generic descr_call and manipulate the __args__ manually.
The test done by 8823f6116665 shows that the JIT generates the very same
code.

By doing this, we probably loose some performance without JIT, because we
don't go through fastcall_{0,1} for the NOARGS and O cases. However, the
advantage is that this approach should work out of the box also for
W_PyCMethodObject.