fix test_sampling_inside_callback by restarting sampling inside new_stacklet_callback, after it was stopped inside switch().
This fix is a bit obscure because one pairt of start/stop is inside
rstacklet.switch, while the other is in interp_continuation.  OTOH, if we do
the fix inside rstacklet we need to replicate it for every GcRootFinder, which
is also obscure and a nightmare to test.  Note sure what is the least ugly :(