Untranslated app-level tests without py or _pytest
This uses a highly modified version of the ast-rewriting in order to
be able to run the app-level tests in untranslated mode without having
to import the py or pytest libraries.  The idea of this is that it
speeds up the test run as not as much has to be imported upfront.  But
not good enough yet.