(arigato, remi) cleanup test and make the bridge not contain a call_pure
By preventing to share descrs with the GUARD_NOT_INVALIDATED coming from the
elidable_compatible decorator, a later guard will not share the descr and
therefore attach the bridge where it failed itself. With that, we don't start
the trace at the GUARD_NOT_INVALIDATED, which is right after a guard_compatible.
(Previously that meant that the bridge didn't contain all the information to
remove a call_pure)