fix issue 2777:
before this commit, the following problem could occur: when the tags of the
opencoder overflow, it raises SwitchToBlackHole in the middle of a jitcode
opcode. Potentially, in the middle of a guard, *before* the guard changed the
pc on the MIFrame! That way, the blackhole interpreter would continue in the
wrong branch :-(.

Fix this by changing the interface slightly: opencoder.Trace.record will now
just not record anything, if the tags overflow. Instead, it will set a flag on
itself, which the metatracer needs to check after every opcode, to find out
whether to stop tracing.