Add rarithmetic.mulmod(). Use it to improve the performance of
 pow(a, b, c)

if c is a large integer that still fits inside a W_IntObject.

Note that this means that pow(a, b, c) now returns an int more
systematically, and not a long, if all arguments are ints. I hope this
is not a problem. Previously we would return sometimes an int and
sometimes a long in this situation, based on the internal algorithm,
like CPython.