CPython issue #38106
On OS X, call pthread_cond_signal() before pthread_mutex_unlock().
This should not make a difference, but does if another thread notices
that the lock is released and frees the lock---this is possible
at any time after the pthread_mutex_unlock().  We can't rely on the
condition variable 'lock->lock_released' to remain in existence.

This can't occur on top of Python lock objects because of the GC,
but can occur by direct calls to the C API functions via cpyext.