(antocuni, arigo) WIP implementing test_exception
this mostly works already. However, if you raise an exception which is raised
inside an llhelper callback, ll2ctypes raises it and ctypes prints it to
stderr (by invoking sys.excepthook). After the printing, it is correctly
propagated and thus the test passes.

Next commits will try to convince ll2ctypes to do the right thing and just
propagate the exception