redo 8fcba04d307f: don't generate tons and tons of guard_not_invalidated while tracing
one benchmark had 43% of all guards be guard_not_invalidated

However, this needs an adaptation in the optimizer, to remove
guard_not_invalidated less aggressively:

we need one guard_not_invalidated after every call that can invalidate
something. This is independent to whether the quasiimmut_field op is removed or
not! The tracer will only trace one guard_not_invalidated after each call, so
even if the first quasiimmut_field is removed, the second one might not be and
could rely on the presence of an earlier guard_not_invalidated. This might
under rare circumstances leave a extra guard_not_invalidated in the trace! But
guard_not_invalidated is cheap, it emits no instructions and its only cost is
the size of the resume data. Therfore that is still a better tradeoff than
capturing resume data for every quasiimmut_field in the front end *all the
time*