Check whether stream is closed in BufferedMixin.truncate_w().
Before this, the method was failing later on a closed stream when accessing the
underlying raw stream. It is more consistent and robust to check for a closed
stream earlier.

Starting with the next changeset, closing the stream will set the buffer to
None. Given that the buffer is potentially accessed by truncate_w(), it is
a very good idea to check that the stream is not closed, although it is hard
to make truncate_w() actually access the buffer when the stream is closed
(because closing it will flush the unwritten data from the buffer).