Fix translation, which was broken by 9a8d23dd614c.
Previously, this hint was not needed because it was already included by
llexternal('ctx_Arg_Parse', ...). But 9a8d23dd614c killed the llexternal, with
the consequence that the eci was no longer seen by the translator and not
included in the final makefile. By adding this, we make sure that it's
included again.