Use timespec on macOS too.
It's present, the fields are just named slightly differently.

I think this fixes all but one test in rlib.rposix_stat, which
previously were failing.

The last one just needs some bifurcation due again to the
differing names in the timespec struct.

More broadly, the bug here is that this code has different
behavior on PyPy and CPython + macOS:

    touch foo
    for each in pypy3 python3.6; do
        echo $each; $each -c 'import os; print(os.stat("foo").st_mtime)'
    done

where on PyPy the *time attributes only give integer precision.

And yeah this actually came up -- a test I wrote for other code failed
on PyPy because the mtime of a file was changing by less than a second
of precision.