Only set CC variable if otherwise unspecified
This probing logic tests for the existence of `gcc` and sets the value of
the 'CC' variable based on the results of that test, but many build pipelines
already define a value for 'CC' that they don't want overridden. Change this
probing logic to only set 'CC' if the environment variable doesn't already have
a specified value.