sync httplib2 and urllib2 to cpython 2.7 git with security backports
Sync httplib and urllib2 stdlib modules and the respective tests
to the current state of CPython 2.7 git (EOL-ed) + two patches
backported from 3.6 that are present in the Gentoo patchset.  This has
only trivial changes compared to backporting the four relevant patches
and should make it easier to apply a final stdlib update post-release.

The Gentoo patches can be found as the two top patches on
https://gitweb.gentoo.org/fork/cpython.git/log/?h=gentoo-2.7.18-r2