KIll W_CPyStaticData.
In the early days, it was needed because we recived HPyMethodDef[] and had to
malloc() PyMethodDef[] on the fly, so this was needed to handle the ownership
of the malloced memory.

Nowadays, the HPy API is written in a way that we receive PyMethodDef[]
direcly, and thus it's the job of the caller to manage its lifetime (and
usually, they are static data anyay).