Refactor interp_slots.py
The autogen approach works for simple cases, but it quickly shows its limit
for more complex ones. The problem is that even in CPython the correspondance
between the various *func typedef and wrap_* functions is strong but not
complete.

So, it seems better and easier to follow CPython as close as possible: instead
of autogenerating the W_SlotWrapper from the HPyFunc_* typedef, we write them
manually, following the very same name convention as CPython.

In this commit, kill autogen_interp_slots and move/rename the exiting wrappers
inside interp_slot.py