Fix a bug in objspace.fake.checkmodule which was introduced 5 years ago by e7c6c1c7
The call to module.init() causes module.startup() to be called at
translation-time and sets module.startup_called = True. The net effect is that
we freeze values which are supposed to be computed at runtime, and this
e.g. causes problems to HPy.

The original commit added the call to module.init() to be able to ztranslate
_multiprocessing: fix it differently by moving the relevant code to startup().