WIP: start a refactoring to kill the old "global" API handles.{new,deref,...}, and replace it with state.handles.{new,deref,...}.
The idea is that W_ExtensionFunction and W_SlotWrapper will need to use a
different HandleManager when we are using the debug mode, so this is a first
step in that direction.

I am going to complete the refactoring in few steps to keep the diff
maneageable. This commit contains:

1. the removal of the old API
2. a blind query/replace to substitute handles.xxx(space, ...) with state.handles.xxx(...)

The code is still broken though, will be fixed in the next commits