Fix and run ./update_vedored.sh.
There has been a refactoring of how versions are handled on the git side. The
new situation is the following:

- git commit facdb01 on master is tagged as 0.0.2rc1

- there is a git branch called release/0.0.2 which branched from there

This pypy branch is now following the corresponding hpy branch. The content of
the files is still the same, so this run of ./update_vendored.sh just changes
the computed version number.