fix pyparser/test/test_pyparse.py::TestPythonParserRevDB::test_encoding
it was failing because the malloc leak finder was pointing out that the buffer
of a buffered reader was actually never released! RevDB is special, because due
to the split address space it is actually using a RawByteBuffer instead of a
ByteBuffer.

fix this by just setting the reference to the buffer to None when closing the
buffered reader