#3673: I am not 100% sure this fixes the remaining trace hook problems, but it has a pretty decent chance. approach is a bit different than CPython, but CPython's approach has problems (broken in 3.9) and is very complicated.
the main problem is that for except Cls as name:
there is an artificial try: finally: block like this:

except Cls as name:
    try:
        ... # user-written code
    finally:
        name = None; del name

we never want to trace that finally, because it has the line number of the last
line in the user-written content of the except block. what I implement here is
to add a NOP in the artificial finally, and use that to detect this case and
not trigger a line trace event then.