Let tp_iternext raise StopIteration more selectively
First, we can deal with the situation of handling an exception or not from the
same wrapped function, using a variable to control whether StopIterations should
be raised or swallowed. Then, the value of such variable is calculated before
the wrapped function is created, and is calculated from a hardcoded list of
types that are known to raise StopIteration in their tp_iternext
implementations.