a different approach to the action dispatcher to fix #4899
the problem is that making sure the ActionFlag.fire method does not
collect is not enough. it must also not change any pointer fields that
would trigger the write barrier to be called (unfortunately we have no
annotation for that).

the bug in 4899 is an invariant in the GC being violated by a GC hook
action being inserted into the linked list of actions.

the new approach is to have a list of all actions and a bitfield that
signifies if an action has fired. that way we don't need to put an
action into a linked list at all if it fires, instead we just set the
corresponding bit.