a test for the bug in #4925
basically if you do things in just the wrong order, then you can indeed
get a black object pointing to a white (prebuilt) object.

so far this is only reproducing the crashing assertion. I think with
some more careful manipulation it should be possible to also get
the gc to collect an object that is still in use with this approach
(which explains the memory corruption problems that some people are
seeing).