fix the failing test
basically if you do things in just the wrong order with prebuilt
objects, then you can get a violation of the invariant that black
objects must only point to grey ones. fix this by making prebuilt
objects grey when they get mutated for the first time, if we are in the
marking phase.