try to fix bug in writebarrier_before_copy
when we are marking, we must not inspect the state of the
GCFLAG_TRACK_YOUNG_PTRS of the source array in writebarrier_before_copy.
the reason for that is that the write barrier is not only responsible
for identifying the old objects that point to young objects. it is
crucially also responsible for identifying the old black objects that
were modified at all since the last collection, to make sure that they
can be turned gray in the next minor collection and get traced again as
part of the incremental marking