Fix LTO issue with pypysig_counter
If `limits.h` isn't included, we get two values in use for `PATH_MAX` -
one is the fallback defined by `signals.h`, and the other is the one
that `signals.c` sees when it includes `limits.h`. This is a problem
with LTO if another object included limits.h and picked up a different
definition of `pypysig_counter`, showing up as -Wlto-type-mismatch.

Fix that by including `limits.h` unconditionally. signals.c already
does that so it should be fine.

Closes: https://github.com/pypy/pypy/issues/5199