gh-5212: fix another corner case in unroll jit optimization
in OptUnroll.inline_short_preamble there is a careful fixpoint to keep
checking whether inlining the short preamble added new short preamble
ops. The fixpoint failed to check whether the inlining leads to more
short preamble label args, which made the number of arguments in the
label and the jump operation go out of sync.