Test the compiled lexer that was ignored before, by mistake
This is probably due to a typo that became clear after cleaning the code and looking at the test using intellij. Intellij told me that the result from compiling the lexer function was never used. The fact that the assertions after the compilation also were identical to the one above the compilation call was hint that something was wrong.